+++ src/utils/gen.py
@@ -401,6 +401,30 @@ def get_pos() -> tuple[int, int] | None:
 #############
 ### INPUT ###
 #############
+# Source - https://stackoverflow.com/a/17939632
+# Posted by Vitja Makarov
+# Retrieved 2026-06-28, License - CC BY-SA 3.0
+def setup_term(fd, when=termios.TCSAFLUSH):
+    mode = termios.tcgetattr(fd)
+    mode[tty.LFLAG] = mode[tty.LFLAG] & ~(termios.ECHO | termios.ICANON)
+    termios.tcsetattr(fd, when, mode)
+
+
+def getch(timeout: float | None = None) -> str | None:
+    fd = sys.stdin.fileno()
+    old_settings = termios.tcgetattr(fd)
+    try:
+        setup_term(fd)
+        try:
+            rw, wl, xl = select.select([fd], [], [], timeout)
+        except select.error:
+            return None
+        if rw:
+            return sys.stdin.read(1)
+    finally:
+        termios.tcsetattr(fd, termios.TCSADRAIN, old_settings)
+
+
 class InpHdlr:
     def __init__(self) -> None:
         self.fd = sys.stdin.fileno()